requirements.txt
Flask==3.0.3
gunicorn==23.0.0
transformers==4.44.2
torch==2.4.1
Pillow==10.4.0
timm